+++ src/Controller/ProprietaireController.php
@@ -6,6 +6,7 @@ use App\Entity\Proprietaire;
 use App\Form\ProprietaireType;
 use App\Repository\LocaliteRepository;
 use App\Repository\ProprietaireRepository;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -15,8 +16,13 @@ use Symfony\Component\Routing\Annotation\Route;
 class ProprietaireController extends AbstractController
 {
     #[Route('/', name: 'app_proprietaire_index', methods: ['GET', 'POST'])]
-    public function index(Request $request, ProprietaireRepository $proprietaireRepository, LocaliteRepository $localiteRepository): Response
+    public function index(Request $request, PaginatorInterface $pgn, ProprietaireRepository $proprietaireRepository, LocaliteRepository $localiteRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         // Définition en session du module en cours
         $request->getSession()->set('menu', 'proprietaire');
         $request->getSession()->set('sousmenu', '');
@@ -32,9 +38,11 @@ class ProprietaireController extends AbstractController
         }
 
         $les_proprietaire = $proprietaireRepository->findByRestr($val_rech, $val_filtre, $orderBy, $page);
+        $list=$les_proprietaire;
+        $list = $pgn->paginate($list, $request->query->getInt('page', 1), 20);
 
         return $this->render('proprietaire/index.html.twig', [
-            'les_proprietaire' => $les_proprietaire,
+            'les_proprietaire' => $list,
             'val_rech' => $val_rech,
 
             'les_localite'=> $les_localite,
@@ -49,6 +57,11 @@ class ProprietaireController extends AbstractController
     #[Route('/add', name: 'app_proprietaire_add', methods: ['GET', 'POST'])]
     public function new(Request $request, ProprietaireRepository $proprietaireRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $proprietaire = new Proprietaire();
         $form = $this->createForm(ProprietaireType::class, $proprietaire);
         $form->handleRequest($request);
@@ -67,6 +80,11 @@ class ProprietaireController extends AbstractController
     #[Route('/{id}', name: 'app_proprietaire_show', methods: ['GET'])]
     public function show(Proprietaire $proprietaire, ProprietaireRepository $proprietaireRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         return $this->render('proprietaire/show.html.twig', [
             'les_proprietaire' => $proprietaireRepository->findAll(),
             'proprietaire' => $proprietaire,
@@ -76,6 +94,11 @@ class ProprietaireController extends AbstractController
     #[Route('/{id}/edit', name: 'app_proprietaire_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Proprietaire $proprietaire, ProprietaireRepository $proprietaireRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $form = $this->createForm(ProprietaireType::class, $proprietaire);
         $form->handleRequest($request);
 
@@ -94,6 +117,11 @@ class ProprietaireController extends AbstractController
     #[Route('/{id}', name: 'app_proprietaire_delete', methods: ['POST'])]
     public function delete(Request $request, Proprietaire $proprietaire, ProprietaireRepository $proprietaireRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$proprietaire->getId(), $request->request->get('_token'))) {
             $proprietaireRepository->remove($proprietaire);
         }
