+++ src/Controller/TypeConstructionController.php
@@ -16,6 +16,11 @@ class TypeConstructionController extends AbstractController
     #[Route('/', name: 'app_type_construction_index', methods: ['GET'])]
     public function index(TypeConstructionRepository $typeConstructionRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         return $this->render('type_construction/index.html.twig', [
             'type_constructions' => $typeConstructionRepository->findAll(),
         ]);
@@ -24,6 +29,11 @@ class TypeConstructionController extends AbstractController
     #[Route('/add', name: 'app_type_construction_add', methods: ['GET', 'POST'])]
     public function new(Request $request, TypeConstructionRepository $typeConstructionRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $typeConstruction = new TypeConstruction();
         $form = $this->createForm(TypeConstruction1Type::class, $typeConstruction);
         $form->handleRequest($request);
@@ -42,6 +52,11 @@ class TypeConstructionController extends AbstractController
     #[Route('/{id}', name: 'app_type_construction_show', methods: ['GET'])]
     public function show(TypeConstruction $typeConstruction): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         return $this->render('type_construction/show.html.twig', [
             'type_construction' => $typeConstruction,
         ]);
@@ -50,6 +65,11 @@ class TypeConstructionController extends AbstractController
     #[Route('/{id}/edit', name: 'app_type_construction_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, TypeConstruction $typeConstruction, TypeConstructionRepository $typeConstructionRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         $form = $this->createForm(TypeConstruction1Type::class, $typeConstruction);
         $form->handleRequest($request);
 
@@ -67,6 +87,11 @@ class TypeConstructionController extends AbstractController
     #[Route('/{id}', name: 'app_type_construction_delete', methods: ['POST'])]
     public function delete(Request $request, TypeConstruction $typeConstruction, TypeConstructionRepository $typeConstructionRepository): Response
     {
+        // Redirection vers page login si session inexistante !!!
+        if(!$this->getUser()) {
+            return $this->redirectToRoute('app_logout', [], Response::HTTP_SEE_OTHER);
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$typeConstruction->getId(), $request->request->get('_token'))) {
             $typeConstructionRepository->remove($typeConstruction);
         }
